bq_sql_function/requirements.txt
functions-framework==3.*
flask
google-cloud-bigquery
vertexai